+++ CHANGELOG.md
@@ -757,4 +757,4 @@ Fixed:
 [0.68.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.68.0
 [1.0.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v1.0.0
 [1.1.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v1.1.0
-[Unreleased]:https://github.com/yagop/node-telegram-bot-api/compare/v1.0.0...master
+[Unreleased]:https://github.com/yagop/node-telegram-bot-api/compare/v1.1.0...master